Fix RCRef for payloads containing GC pointers.
Was missing a GC.addRange call to be safe to use for payloads with GC references.

Also, this fixes a bug in the reset() method, where the m_context pointer was not reset to null - note that this bug never manifested, because the reset() method was not used anywhere outside of the destructor.